move pilosa/test/attr.go into pilosa/attr_test.go
