+++ attr_test.go
@@ -15,15 +15,20 @@
 package pilosa_test
 
 import (
+	"io/ioutil"
+	"os"
 	"reflect"
+	"runtime"
+	"sync"
 	"testing"
 
-	"github.com/pilosa/pilosa/test"
+	"github.com/pilosa/pilosa"
+	"github.com/pilosa/pilosa/boltdb"
 )
 
 // Ensure database can set and retrieve column attributes.
 func TestAttrStore_Attrs(t *testing.T) {
-	s := test.MustOpenAttrStore()
+	s := MustOpenAttrStore()
 	defer s.Close()
 
 	// Set attributes.
@@ -52,7 +57,7 @@ func TestAttrStore_Attrs(t *testing.T) {
 
 // Ensure database returns a non-nil empty map if unset.
 func TestAttrStore_Attrs_Empty(t *testing.T) {
-	s := test.MustOpenAttrStore()
+	s := MustOpenAttrStore()
 	defer s.Close()
 
 	if m, err := s.Attrs(100); err != nil {
@@ -64,7 +69,7 @@ func TestAttrStore_Attrs_Empty(t *testing.T) {
 
 // Ensure database can unset attributes if explicitly set to nil.
 func TestAttrStore_Attrs_Unset(t *testing.T) {
-	s := test.MustOpenAttrStore()
+	s := MustOpenAttrStore()
 	defer s.Close()
 
 	// Set attributes.
@@ -84,7 +89,7 @@ func TestAttrStore_Attrs_Unset(t *testing.T) {
 
 // Ensure attribute block checksums can be returned.
 func TestAttrStore_Blocks(t *testing.T) {
-	s := test.MustOpenAttrStore()
+	s := MustOpenAttrStore()
 	defer s.Close()
 
 	// Set attributes.
@@ -122,4 +127,68 @@ func TestAttrStore_Blocks(t *testing.T) {
 	} else if !reflect.DeepEqual(blks0[2], blks1[2]) {
 		t.Fatalf("block 2 mismatch: %#v != %#v", blks0[2], blks1[2])
 	}
+}
+
+// AttrStore represents a test wrapper for pilosa.AttrStore.
+type AttrStore struct {
+	pilosa.AttrStore
+}
+
+// NewAttrStore returns a new instance of AttrStore.
+func NewAttrStore(string) pilosa.AttrStore {
+	f, err := ioutil.TempFile("", "pilosa-attr-")
+	if err != nil {
+		panic(err)
+	}
+	f.Close()
+	os.Remove(f.Name())
+
+	return &AttrStore{boltdb.NewAttrStore(f.Name())}
+}
+
+func BenchmarkAttrStore_Duplicate(b *testing.B) {
+	s := MustOpenAttrStore()
+	defer s.Close()
+
+	// Set attributes.
+	const n = 5
+	for i := 0; i < n; i++ {
+		if err := s.SetAttrs(uint64(i), map[string]interface{}{"A": 100, "B": "foo", "C": true, "D": 100.2}); err != nil {
+			b.Fatal(err)
+		}
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	// Update attributes with an existing subset.
+	cpuN := runtime.GOMAXPROCS(0)
+	var wg sync.WaitGroup
+	for i := 0; i < cpuN; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			for j := 0; j < b.N/cpuN; j++ {
+				if err := s.SetAttrs(uint64(j%n), map[string]interface{}{"A": int64(100), "B": "foo", "D": 100.2}); err != nil {
+					b.Fatal(err)
+				}
+			}
+		}()
+	}
+	wg.Wait()
+}
+
+// MustOpenAttrStore returns a new, opened attribute store at a temporary path. Panic on error.
+func MustOpenAttrStore() pilosa.AttrStore {
+	s := NewAttrStore("")
+	if err := s.Open(); err != nil {
+		panic(err)
+	}
+	return s
+}
+
+// Close closes the database and removes the underlying data.
+func (s *AttrStore) Close() error {
+	defer os.RemoveAll(s.Path())
+	return s.AttrStore.Close()
 }
